+++ test/test_identification.rb
@@ -15,7 +15,6 @@ class TestFakerIdentification < Test::Unit::TestCase
   end
 
   def test_drivers_license
-    ###-###-##-###-#
     drivers_license_regex = /\A[A-Z]\d{3}-\d{3}-\d{2}-\d{3}-\d{1}\z/
     assert_match(drivers_license_regex, @tester.drivers_license)
   end
@@ -29,7 +28,18 @@ class TestFakerIdentification < Test::Unit::TestCase
   end
 
   def test_ethnicity
-    ethnicity_regex = %r{\A(African American|Asian/Pacific Islander|Caucasian|Hispanic|Native American|Multiracial|Other|Prefer not to respond)\z}
+    ethnicity_regex = %r{
+      \A(
+        African\ American|
+        Asian/Pacific\ Islander|
+        Caucasian|
+        Hispanic|
+        Native\ American|
+        Multiracial|
+        Other|
+        Prefer\ not\ to\ respond
+      )\z
+    }x
     assert_match(ethnicity_regex, @tester.ethnicity)
   end
 end